refactor(app): sort saveBlob de la feature preuves vers les utilitaires
`saveBlob` sauvegarde un blob sur le disque : il ne sait rien des preuves ni
des referentiels. Il vivait dans `referentiels/preuves/Bibliotheque/` parce que
les preuves en ont ete le premier usage.

Ses quatorze appelants se repartissent sur six domaines — graphiques,
trajectoire, indicateurs, plans, demarches, referentiels — et
`utils/useDownloadFile.ts` l'importait depuis une feature, inversant la
dependance.

Renomme au passage en kebab-case, convention des fichiers du front. Aucune
ligne de logique modifiee.
