+++ apps/app/src/plans/plans/show-plan/data/use-export-plan.ts
@@ -1,5 +1,5 @@
 import { appLabels } from '@/app/labels/catalog';
-import { saveBlob } from '@/app/referentiels/preuves/Bibliotheque/saveBlob';
+import { saveBlob } from '@/app/utils/save-blob';
 import { useApiClient } from '@/app/utils/use-api-client';
 import { useMutation } from '@tanstack/react-query';
 import { useCollectiviteId } from '@tet/api/collectivites';
