+++ apps/app/src/referentiels/preuves/data/use-download-document.ts
@@ -1,4 +1,4 @@
-import { saveBlob } from '@/app/referentiels/preuves/Bibliotheque/saveBlob';
+import { saveBlob } from '@/app/utils/save-blob';
 import { DOWNLOAD_FILE_MUTATION_OPTIONS } from '@/app/utils/toast/download-file-mutation-options';
 import { useMutation, type UseMutationResult } from '@tanstack/react-query';
 import { useTRPC } from '@tet/api';
